un poco de todo

+++ SpaceInvaderII/src/tp/p1/logic/objects/RegularShip.java
@@ -17,11 +17,12 @@ public class RegularShip extends AlienShip{
 	}
 	
 	public void computerAction() {
-		/*if(IExecuteRandomActions.canGenerateRandomExplosiv(game)) {
-			ExplosivShip explosiv = new ExplosivShip(this.game, super.x, super.y, this.life, 1);
+		if(IExecuteRandomActions.canGenerateRandomExplosiv(game)) {
+			ExplosivShip explosiv = new ExplosivShip(this.game, super.x, super.y, this.life,super.getcyclesToMove());
 			this.game.addObject(explosiv);
 			this.game.deleteOnlist(this);
-		}*/
+			super.setREMAINING_ALIENS();
+		}
 	}
 	
 	public void hurt() {
